Add authenticate and load functions

+++ src/main.rs
@@ -1,3 +1,4 @@
+extern crate sha2;
 use sha2::{Sha256, Digest};
 use std::fmt;
 use std::fs::File;
@@ -60,9 +61,9 @@ impl fmt::Debug for Usuario{
     }
 }
 impl Usuario {
-    pub fn new(nome: String, senha: String) -> Usuario {
+    pub fn new(nome: &String, senha: &String) -> Usuario {
 
-	Usuario{nome, senha:sha256(senha)}
+	Usuario{nome: nome.to_string(), senha: sha256(senha.to_string())}
 
         // Não se esqueça de armazenar a senha do usuário como sha256
     }
@@ -77,8 +78,8 @@ impl <'a> Cadastro<'a> {
         Cadastro{ usuarios: vec![], arquivo_usuarios }
     }
 
-    pub fn adiciona_usuario(&mut self, nome: &String, senha: &String) {
-	&self.usuarios.push(Usuario{nome:nome.to_string(), senha:senha.to_string()});
+    pub fn adiciona_usuario(&mut self, user: Usuario) {
+	    self.usuarios.push(user);
     }
 
     fn usuario_existe(&self, nome: &String) -> bool {
@@ -88,20 +89,20 @@ impl <'a> Cadastro<'a> {
         return false
     }
 
-    fn salvar(&self, mut file: &File, i:&isize) -> std::io::Result<()> {
-        let f = match File::options().append(true).open(&self.arquivo_usuarios){
+    fn salvar(&self, i:&isize) -> std::io::Result<()> {
+        let mut f = match File::options().append(true).open(&self.arquivo_usuarios){
             Err(e) => panic!("Erro ao abrir o arquivo pelo motivo {}", e),
-            Ok(f) => {&file}
+            Ok(f) => {f}
         };
 
         let conteudo: &Usuario = &self.usuarios[*i as usize];
 
-        match writeln!(&mut file, "{:?}",conteudo){
+        match writeln!(&mut f, "{:?}",conteudo){
             Err(e) => panic!("Erro ao escrever no arquivo pelo motivo {}", e),
             Ok(()) => {}
         }
             
-        match file.flush(){
+        match f.flush(){
             Err(e) => panic!("Erro ao salvar o arquivo pelo motivo {}", e),
             Ok(()) => {}
         }
@@ -110,8 +111,36 @@ impl <'a> Cadastro<'a> {
 
     }
 
-    fn autenticar(&self, nome: String, senha: String) -> bool {
-        todo!("Implementar a autenticação de usuários!");
+    fn autenticar(&mut self, nome: &String, senha: &String) -> bool {
+        if self.usuario_existe(&nome) {
+            if self.arquivo_usuarios.exists() {
+                let _ = self.carregar();
+
+                let mut arquivo: File = match File::open(self.arquivo_usuarios) {
+                    Err(e) => panic!("Erro ao abrir o arquivo pelo motivo {}", e),
+                    Ok(arquivo) => {arquivo}
+                };
+    
+                let mut conteudo: String = String::new();
+    
+                match arquivo.read_to_string(&mut conteudo) {
+                    Err(e) => panic!("Erro ao abrir o arquivo pelo motivo {}", e),
+                    Ok(_) => println!("")
+                };
+    
+                let user: String = format!(r#"nome: "{}", senha: "{}""#, nome, senha);
+    
+                if conteudo.contains(&user) {
+                    println!("Usuário autenticado!");
+                    return true;
+                } else {
+                    println!("Falha ao autenticar o usuário! Verifique suas credenciais.");
+                    return false;
+                }
+            }
+        }
+        println!("O usuário não existe.");
+        return false;
     }
 
     fn retorna_indice_usuario(&self, nome: &String) -> isize {
@@ -128,7 +157,31 @@ impl <'a> Cadastro<'a> {
     }
 
     fn carregar(&mut self) -> std::io::Result<()>{
-       todo!("Implementar o carregamento de arquivos!");
+       let mut arquivo: File = match File::open(self.arquivo_usuarios) {
+            Err(e) => panic!("Erro ao abrir o arquivo pelo motivo {}", e),
+            Ok(arquivo) => {arquivo}
+        };
+
+        let mut conteudo: String = String::new();
+
+        match arquivo.read_to_string(&mut conteudo) {
+            Err(e) => panic!("Erro ao abrir o arquivo pelo motivo {}", e),
+            Ok(_) => println!("")
+        };
+
+        for line in conteudo.lines() {
+            let line_str = line.to_string();
+            let split_line: Vec<&str> = line_str.split('"').collect();
+
+            if split_line.len() > 1 {
+                let nome: String = String::from(split_line[1]);
+                let senha: String = String::from(split_line[3]);
+
+                let user: Usuario = Usuario::new(&nome, &senha);
+
+                self.adiciona_usuario(user);
+            }
+        }
 
        Ok(()) // Não apague esta linha
     }
@@ -136,28 +189,63 @@ impl <'a> Cadastro<'a> {
 
 fn main() {
         let path = Path::new("foo.txt");
-        let mut f = match File::create(path){
-            Err(e) => panic!("Erro ao ler o arquivo pelo motivo {}", e),
-            Ok(f) => {f}
-        };
+        if !path.exists() {
+            let mut _f = match File::create(path){
+                Err(e) => panic!("Erro ao ler o arquivo pelo motivo {}", e),
+                Ok(f) => {f}
+            };
+        }
+        
         let mut cadastro: Cadastro = Cadastro::new(&path);
-        while true{
-            let mut nome: String = String::from("");
-            let mut senha: String = String::from("");
-            println!("Digite o seu username: ");
-            io::stdin().read_line(&mut nome).expect("Erro ao ler o username");
-            println!("Digite a sua senha: ");
-            io::stdin().read_line(&mut senha).expect("Erro ao ler a senha");
-            let nome_s: String = nome.trim().parse().expect("Erro ao transformar o username");
-            let senha_s:String = nome.trim().parse().expect("Erro ao transformar a senha");
-            let user: Usuario = Usuario::new(nome_s, senha_s);
-            if cadastro.usuario_existe(&user.nome).eq(&false){
-                cadastro.adiciona_usuario(&user.nome, &user.senha);
-                let i = cadastro.retorna_indice_usuario(&user.nome);
-                println!("Usuario {} cadastrado com sucesso", user.nome);
-                cadastro.salvar(&f, &i);
-            }else{
-                println!("Usuario {} já existe", user.nome);
+        let _ = cadastro.carregar();
+        println!("{:?}", &cadastro.usuarios);
+        loop {
+            let mut menu: String = String::new();
+
+            println!("Digite a ação que você deseja:\n 1 - Autenticar usuário.\n 2 - Cadastrar usuário.\n 3 - Sair.\n");
+            io::stdin().read_line(&mut menu).expect("Erro ao ler a entrada");
+
+            let menu_s: String = menu.trim().parse().expect("Erro ao converter a opção selecionada");
+
+            if menu_s == "1".to_string() {
+                let mut nome: String = String::new();
+                let mut senha: String = String::new();
+
+                println!("Digite o seu username: ");
+                io::stdin().read_line(&mut nome).expect("Erro ao ler o username");
+                println!("Digite a sua senha: ");
+                io::stdin().read_line(&mut senha).expect("Erro ao ler a senha");
+
+                let nome_s: String = nome.trim().parse().expect("Erro ao transformar o username");
+                let senha_s:String = senha.trim().parse().expect("Erro ao transformar a senha");
+
+                cadastro.autenticar(&nome_s, &sha256(senha_s));
+            } else if menu_s == "2".to_string() {
+                let mut nome: String = String::new();
+                let mut senha: String = String::new();
+
+                println!("Digite o seu username: ");
+                io::stdin().read_line(&mut nome).expect("Erro ao ler o username");
+                println!("Digite a sua senha: ");
+                io::stdin().read_line(&mut senha).expect("Erro ao ler a senha");
+
+                let nome_s: String = nome.trim().parse().expect("Erro ao transformar o username");
+                let senha_s:String = senha.trim().parse().expect("Erro ao transformar a senha");
+
+                if cadastro.usuario_existe(&nome_s).eq(&false){
+                    let user: Usuario = Usuario::new(&nome_s, &senha_s);
+                    cadastro.adiciona_usuario(user);
+                    let i = cadastro.retorna_indice_usuario(&nome_s);
+                    println!("Usuario {} cadastrado com sucesso", &nome_s);
+                    let _ = cadastro.salvar(&i);
+                } else {
+                    println!("Usuário já existe.");
+                }
+            } else if menu_s == "3".to_string() {
+                println!("Fechando o programa.");
+                break;
+            } else {
+                println!("Comando inválido!");
             }
         }
 }
